Chore: Integrate network functions
